pyproject.toml
[tool.poetry]
name = "langChainDemo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain-openai = "^0.0.8"
pirate-speak = {path = "packages/pirate-speak", develop = true}
pydantic = "1.10.13"
langchainhub = "^0.1.15"
sentence-transformers = "^2.5.1"
qdrant-client = "^1.8.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
